dbapi/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
ngrok==0.12.1
numpy==1.26.3
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
zipp==3.17.0
gunicorn
